เพิ่มการยกเลิกการเป็นสมาชิกกองทุน  แต่ยังเก็บข้อมูลไว้

+++ application/views/listMember.php
@@ -67,7 +67,7 @@
       <td align="center" valign="middle" nowrap="nowrap" style="font-size: 12px"><a href="<?php echo base_url();?>index.php/admin/editMemberView/<?php echo $m['memberId'];?>" class="editMember fancybox.iframe">
       <img class="iconAction" src="<?php echo base_url();?>img/editIcon.png" width="25px" height="25px"  style="margin-bottom:-8px;">
       </a></td>
-      <td align="center" valign="middle" nowrap="nowrap" style="font-size: 12px"><a href="<?php echo base_url();?>index.php/admin/deleteMember/<?php echo $m['memberId'];?>" class="deleteMember fancybox.iframe">
+      <td align="center" valign="middle" nowrap="nowrap" style="font-size: 12px"><a onClick="return confirme()" href="<?php echo base_url();?>index.php/admin/deleteMember/<?php echo $m['memberId'];?>" class="deleteMember fancybox.iframe">
       <img class="iconAction" src="<?php echo base_url();?>img/deleteIcon.png" width="25px" height="25px"  style="margin-bottom:-8px;">
       </a></td>
     </tr>
